ref(upstream): Update the public interface
