+++ relay-server/src/endpoints/forward.rs
@@ -7,9 +7,9 @@ use std::borrow::Cow;
 use std::fmt;
 
 use ::actix::prelude::*;
-use actix_web::error::ResponseError;
+use actix_web::error::{ParseError, ResponseError};
 use actix_web::http::header::{self, HeaderName, HeaderValue};
-use actix_web::http::{uri::PathAndQuery, HeaderMap, Method, StatusCode};
+use actix_web::http::{uri::PathAndQuery, HeaderMap, StatusCode};
 use actix_web::{AsyncResponder, Error, HttpMessage, HttpRequest, HttpResponse};
 use bytes::Bytes;
 use futures01::{future, prelude::*, sync::oneshot};
@@ -19,7 +19,9 @@ use relay_common::GlobMatcher;
 use relay_config::Config;
 use relay_log::LogError;
 
-use crate::actors::upstream::{SendRequest, UpstreamRelay, UpstreamRequest, UpstreamRequestError};
+use crate::actors::upstream::{
+    Method, SendRequest, UpstreamRelay, UpstreamRequest, UpstreamRequestError,
+};
 use crate::body::RequestBody;
 use crate::endpoints::statics;
 use crate::extractors::ForwardedFor;
@@ -181,38 +183,38 @@ impl UpstreamRequest for ForwardRequest {
         builder.body(&self.data)
     }
 
-    fn respond(
-        &mut self,
-        result: Result<Response, UpstreamRequestError>,
-    ) -> ResponseFuture<(), ()> {
-        let sender = self.sender.take();
+    // fn respond(
+    //     &mut self,
+    //     result: Result<Response, UpstreamRequestError>,
+    // ) -> ResponseFuture<(), ()> {
+    //     let sender = self.sender.take();
 
-        match result {
-            Ok(response) => {
-                let status = StatusCode::from_u16(response.status().as_u16()).unwrap();
-                let headers = response.clone_headers();
+    //     match result {
+    //         Ok(response) => {
+    //             let status = StatusCode::from_u16(response.status().as_u16()).unwrap();
+    //             let headers = response.clone_headers();
 
-                let future = response
-                    .bytes(self.max_response_size)
-                    .and_then(move |body| Ok((status, headers, body)))
-                    .map_err(UpstreamRequestError::Http)
-                    .then(|result| {
-                        if let Some(sender) = sender {
-                            sender.send(result).ok();
-                        }
-                        Ok(())
-                    });
+    //             let future = response
+    //                 .bytes(self.max_response_size)
+    //                 .and_then(move |body| Ok((status, headers, body)))
+    //                 .map_err(UpstreamRequestError::Http)
+    //                 .then(|result| {
+    //                     if let Some(sender) = sender {
+    //                         sender.send(result).ok();
+    //                     }
+    //                     Ok(())
+    //                 });
 
-                Box::new(future)
-            }
-            Err(e) => {
-                if let Some(sender) = sender {
-                    sender.send(Err(e)).ok();
-                }
-                Box::new(future::err(()))
-            }
-        }
-    }
+    //             Box::new(future)
+    //         }
+    //         Err(e) => {
+    //             if let Some(sender) = sender {
+    //                 sender.send(Err(e)).ok();
+    //             }
+    //             Box::new(future::err(()))
+    //         }
+    //     }
+    // }
 }
 
 /// Implementation of the forward endpoint.
@@ -234,7 +236,10 @@ pub fn forward_upstream(
         .unwrap_or("")
         .to_owned();
 
-    let method = request.method().clone();
+    let Ok(method) = Method::from_bytes(request.method().as_ref().as_bytes()) else {
+        return Box::new(future::err(Error::from(ParseError::Method)))
+    };
+
     let headers = request.headers().clone();
     let forwarded_for = ForwardedFor::from(request);
 
@@ -253,7 +258,7 @@ pub fn forward_upstream(
                 sender: Some(tx),
             };
 
-            UpstreamRelay::from_registry().do_send(SendRequest(forward_request));
+            UpstreamRelay::from_registry().send(SendRequest(forward_request));
 
             rx.map_err(|_| UpstreamRequestError::ChannelClosed)
                 .flatten()
